Moved some classes to registry-commons project
