+++ src/main/java/gov/nasa/pds/harvest/dao/SchemaUtils.java
@@ -2,7 +2,8 @@ package gov.nasa.pds.harvest.dao;
 
 import java.util.Set;
 
-import gov.nasa.pds.harvest.meta.FieldNameCache;
+import gov.nasa.pds.registry.common.meta.FieldNameCache;
+
 
 /**
  * Utility methods to work with elasticsearch schema
