+++ src/main/java/gov/nasa/pds/harvest/meta/ex/AutogenExtractor.java
@@ -12,11 +12,11 @@ import org.w3c.dom.NodeList;
 
 import gov.nasa.pds.harvest.Constants;
 import gov.nasa.pds.harvest.job.Job;
-import gov.nasa.pds.harvest.meta.FieldMap;
-import gov.nasa.pds.harvest.util.xml.NsUtils;
-import gov.nasa.pds.harvest.util.xml.XmlDomUtils;
-import gov.nasa.pds.harvest.util.xml.XmlNamespaces;
+import gov.nasa.pds.registry.common.util.FieldMap;
 import gov.nasa.pds.registry.common.util.date.PdsDateConverter;
+import gov.nasa.pds.registry.common.util.xml.NsUtils;
+import gov.nasa.pds.registry.common.util.xml.XmlDomUtils;
+import gov.nasa.pds.registry.common.util.xml.XmlNamespaces;
 
 
 /**
