+++ src/main/java/gov/nasa/pds/harvest/meta/ex/BundleMetadataExtractor.java
@@ -10,8 +10,8 @@ import org.w3c.dom.Document;
 import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
 
-import gov.nasa.pds.harvest.meta.FieldMap;
-import gov.nasa.pds.harvest.util.xml.XPathUtils;
+import gov.nasa.pds.registry.common.util.FieldMap;
+import gov.nasa.pds.registry.common.util.xml.XPathUtils;
 
 
 /**
