+++ src/main/java/gov/nasa/pds/harvest/meta/ex/InternalReferenceExtractor.java
@@ -7,8 +7,8 @@ import org.w3c.dom.Document;
 import org.w3c.dom.Node;
 import org.w3c.dom.NodeList;
 
-import gov.nasa.pds.harvest.meta.FieldMapSet;
-import gov.nasa.pds.harvest.util.xml.XPathUtils;
+import gov.nasa.pds.registry.common.util.FieldMapSet;
+import gov.nasa.pds.registry.common.util.xml.XPathUtils;
 
 
 /**
